+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Polyverse Boost ReST API (Backend)
 
 # Release Notes
 
-## Version 0.9.3: January 22nd, 2024
+## Version 0.9.3: January 24nd, 2024
 
 ### New Features
 - Added REST API /api/status (GET) to get current API version and status and stage
@@ -13,6 +13,7 @@ Polyverse Boost ReST API (Backend)
 ### Enhancements
 - Disabled unsigned authentication headers (x-user-account) - only signed JWT is supported now
 - Abort early if generation of Architectural Spec by AI is failing frequently or a continuous series of failures
+- Use the actual OpenAI file created timestamp on the File Ids (instead of HTTP file upload time)
 
 ### Bug Fixes
 - N/A